ADD ft. to compose & add an individual img result
Added the function addResult() to add an image result to the results
grid. So far the function takes an url as an argument.

The login button is used to add the same image on each click.

An event listener was added on page loaded, from which other listeners
are added.