ADD Images from an API request showing
At load, the page now shows the 10 images referenced by the response
to the basic request made to the Unsplash API. In the results
grid the images were added to those hardcoded for testing. The images
from Unsplash also show a rudimentary tooltip with image metadata.

The ID of each picture is now included in the corresponding html
element for further manipulation. Also the event listeners for the
image favorite button and for detecting a click on the image to open
the full size versions were added.