REFACTORED generic ft. added for fetching http req
A function that accepts an url and a json payload object was added
to be used for every API request. The async function
fetchHttpRequest(urlString, payloadObj) returns a json object with
the resolved response from the API.

This function requires correctly construction of the url and the
payload object prior to be called.

In next commits similar generic functions will be added for
constructing both the url string with parameters and the payload
object with method and headers fields.

This will add more legibility, modularity and re-usability
to the code.