ADD more OAuth testing and full-size img retrieval
Now when clicking an image thumbnail in the results of a search,
a new browser tab opens with the full size image.

The OAuth experiments continued to a dead end, until the dockerization
and the clear division between frontend and backend is implemented.
At this point, the authentication server responds with POST and should
include the access token, but to actually get it, my redirect uri
must be implemented in the backend in a server with the capability of
extracting the POST data.